fix(crm): 修复商机模块多个问题
1. 修复 count 字段数据类型不匹配（Integer -> BigDecimal）
2. 添加 contactId 字段支持联系人关联
3. 添加 contactName 在商机列表和详情中的返回
4. 修复创建商机时 product_price 字段缺失导致的数据库错误
5. 添加数据库迁移脚本

+++ Server/mitedtsm-module-crm/src/main/java/com/meession/etm/module/crm/controller/admin/business/vo/business/CrmBusinessSaveReqVO.java
@@ -88,7 +88,7 @@ public class CrmBusinessSaveReqVO {
 
         @Schema(description = "产品数量", requiredMode = Schema.RequiredMode.REQUIRED, example = "8911")
         @NotNull(message = "产品数量不能为空")
-        private Integer count;
+        private BigDecimal count;
 
     }
